start refactoring main.development
