+++ app/actions/ControlActions.js
@@ -12,7 +12,7 @@ import { ChangePassphraseRequest, RenameAccountRequest,  RescanRequest,
   SetMaxPriceRelativeRequest, SetVotingAddressRequest, SetPoolAddressRequest, SetPoolFeesRequest,
   SetMaxPerBlockRequest,
 } from "../middleware/walletrpc/api_pb";
-import { getWalletCfg } from "../config.js";
+import { getWalletCfg } from "config";
 
 export const GETNEXTADDRESS_ATTEMPT = "GETNEXTADDRESS_ATTEMPT";
 export const GETNEXTADDRESS_FAILED = "GETNEXTADDRESS_FAILED";
