+++ app/config.js
@@ -1,13 +1,18 @@
 import fs from "fs";
-import path from "path";
 import os from "os";
-import { stakePoolInfo } from "./middleware/stakepoolapi";
+import path from "path";
 import Store from "electron-store";
 import ini from "ini";
+import { stakePoolInfo } from "./middleware/stakepoolapi";
+import { getGlobalCfgPath, dcrdCfg, getWalletPath, dcrctlCfg, dcrwalletCfg, appDataDirectory } from "./main_dev/paths";
 
 export function getGlobalCfg() {
   const config = new Store();
   return (config);
+}
+
+export function getWalletCfgPath(testnet, wallet) {
+  return path.resolve(path.join(appDataDirectory(), "wallets", testnet ? "testnet" : "mainnet", wallet));
 }
 
 export function getWalletCfg(testnet, walletPath){
@@ -89,14 +94,6 @@ export function initGlobalCfg() {
   return(config);
 }
 
-export function getGlobalCfgPath() {
-  return path.resolve(appDataDirectory(), "config.json");
-}
-
-export function getWalletCfgPath(testnet, wallet) {
-  return path.resolve(path.join(appDataDirectory(), "wallets", testnet ? "testnet" : "mainnet", wallet));
-}
-
 export function validateGlobalCfgFile() {
   var fileContents;
   try {
@@ -115,34 +112,6 @@ export function validateGlobalCfgFile() {
   }
 
   return null;
-}
-
-// In all the functions below the Windows path is constructed based on
-// os.homedir() rather than using process.env.LOCALAPPDATA because in my tests
-// that was available when using the standalone node but not there when using
-// electron in production mode.
-export function appDataDirectory() {
-  if (os.platform() == "win32") {
-    return path.join(os.homedir(), "AppData", "Local", "Decrediton");
-  } else if (process.platform === "darwin") {
-    return path.join(os.homedir(), "Library","Application Support","decrediton");
-  } else {
-    return path.join(os.homedir(),".config","decrediton");
-  }
-}
-
-export function getDcrdPath() {
-  if (os.platform() == "win32") {
-    return path.join(os.homedir(), "AppData", "Local", "Dcrd");
-  } else if (process.platform === "darwin") {
-    return path.join(os.homedir(), "Library","Application Support","dcrd");
-  } else {
-    return path.join(os.homedir(),".dcrd");
-  }
-}
-
-export function getWalletPath(testnet, walletPath) {
-  return path.join(appDataDirectory(), "wallets", testnet ? "testnet" : "mainnet", walletPath);
 }
 
 export function getWalletCert(certPath) {
@@ -332,15 +301,4 @@ export function newWalletConfigCreation(testnet, walletPath) {
     },
   };
   fs.writeFileSync(dcrwalletCfg(getWalletPath(testnet, walletPath)), ini.stringify(dcrwConf));
-}
-export function dcrctlCfg(configPath) {
-  return path.resolve(configPath, "dcrctl.conf");
-}
-
-export function dcrdCfg(configPath) {
-  return path.resolve(configPath, "dcrd.conf");
-}
-
-export function dcrwalletCfg(configPath) {
-  return path.resolve(configPath, "dcrwallet.conf");
 }
