+++ app/main_dev/paths.js
@@ -1,0 +1,46 @@
+import path from "path";
+import os from "os";
+
+// In all the functions below the Windows path is constructed based on
+// os.homedir() rather than using process.env.LOCALAPPDATA because in my tests
+// that was available when using the standalone node but not there when using
+// electron in production mode.
+export function appDataDirectory() {
+  if (os.platform() == "win32") {
+    return path.join(os.homedir(), "AppData", "Local", "Decrediton");
+  } else if (process.platform === "darwin") {
+    return path.join(os.homedir(), "Library","Application Support","decrediton");
+  } else {
+    return path.join(os.homedir(),".config","decrediton");
+  }
+}
+
+export function getGlobalCfgPath() {
+  return path.resolve(appDataDirectory(), "config.json");
+}
+
+export function getWalletPath(testnet, walletPath) {
+  return path.join(appDataDirectory(), "wallets", testnet ? "testnet" : "mainnet", walletPath);
+}
+
+export function dcrctlCfg(configPath) {
+  return path.resolve(configPath, "dcrctl.conf");
+}
+
+export function dcrdCfg(configPath) {
+  return path.resolve(configPath, "dcrd.conf");
+}
+
+export function dcrwalletCfg(configPath) {
+  return path.resolve(configPath, "dcrwallet.conf");
+}
+
+export function getDcrdPath() {
+  if (os.platform() == "win32") {
+    return path.join(os.homedir(), "AppData", "Local", "Dcrd");
+  } else if (process.platform === "darwin") {
+    return path.join(os.homedir(), "Library","Application Support","dcrd");
+  } else {
+    return path.join(os.homedir(),".dcrd");
+  }
+}
