+++ app/middleware/grpc/client.js
@@ -2,7 +2,8 @@ process.env["GRPC_SSL_CIPHER_SUITES"] = "HIGH+ECDSA";
 
 import grpc from "grpc";
 
-import { getWalletCert, getWalletPath } from "config.js";
+import { getWalletCert } from "config.js";
+import { getWalletPath } from "main_dev/paths.js";
 var services = require("../walletrpc/api_grpc_pb.js");
 
 const getServiceClient = (clientClass) => (isTestNet, walletPath, address, port, cb) => {
